pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-vessel"
version = "0.1.0"
description = "Marine weather intelligence and scheduling CLI"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Weather Vessel Team" }]
dependencies = [
    "httpx>=0.25",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "pytz>=2023.3",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
all = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "black>=23.7",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.5",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
